fix!: rename screw_pitch to screw_lead
I have since learned that the movement caused by a full rotation of the
lead screw is called the lead.
A 1 start 2mm pitch lead screw has a lead of 2mm.
But the SV06 has a 2mm pitch 2 start, with a lead of 4mm.

This caused wrong estimated turns with my example configuration.